+++ dct.h
@@ -7,27 +7,27 @@
 // The last four functions do a discrete cosine transform on the 4 brightnesses
 
 // Return y1 value
-extern double luma_tl(double a, double b, double c, double d);
+extern float luma_tl(float a, float b, float c, float d);
 
 // Return y2 value
-extern double luma_tr(double a, double b, double c, double d);
+extern float luma_tr(float a, float b, float c, float d);
 
 // Retyurn y3 value
-extern double luma_bl(double a, double b, double c, double d);
+extern float luma_bl(float a, float b, float c, float d);
 
 // Return y4 value
-extern double luma_br(double a, double b, double c, double d);
+extern float luma_br(float a, float b, float c, float d);
 
 // Return a value
-extern double coef_a(double y1, double y2, double y3, double y4);
+extern float coef_a(float y1, float y2, float y3, float y4);
 
 // Return b value
-extern double coef_b(double y1, double y2, double y3, double y4);
+extern float coef_b(float y1, float y2, float y3, float y4);
 
 // Return c value
-extern double coef_c(double y1, double y2, double y3, double y4);
+extern float coef_c(float y1, float y2, float y3, float y4);
 
 // Return d value
-extern double coef_d(double y1, double y2, double y3, double y4);
+extern float coef_d(float y1, float y2, float y3, float y4);
 
 #endif
